# apps/web/src/api/generated/openapi.json
{
 "components": {
  "schemas": {
...
   "FeedPreviewResult": {
    "description": "POST /api/v1/feed-preview and POST /api/v1/rsshub/preview.\n\nOnly reliable metadata — no entries, no scraping.",
    "properties": {
     "alreadySubscribed": {
      "title": "Alreadysubscribed",
      "type": "boolean"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "feedUrl": {
      "title": "Feedurl",
      "type": "string"
     },
     "format": {
      "enum": [
       "rss",
       "atom"
      ],
      "title": "Format",
      "type": "string"
     },
     "siteUrl": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Siteurl"
     },
     "title": {
      "title": "Title",
      "type": "string"
     }
    },
    "required": [
     "title",
     "feedUrl",
     "format",
     "alreadySubscribed"
    ],
    "title": "FeedPreviewResult",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "RssHubConfigGroup": {
    "description": "Grouped config items for the Control Center UI.",
    "properties": {
     "id": {
      "title": "Id",
      "type": "string"
     },
     "items": {
      "items": {
       "$ref": "#/components/schemas/RssHubConfigItem"
      },
      "title": "Items",
      "type": "array"
     },
     "label": {
      "title": "Label",
      "type": "string"
     }
    },
    "required": [
     "id",
     "label",
     "items"
    ],
    "title": "RssHubConfigGroup",
    "type": "object"
   },
   "RssHubConfigItem": {
    "description": "One allow-listed RSSHub setting (secrets surface only ``configured``).\n\n``value`` is present only for non-secret items and ``configured`` only\nfor secret items (the route serializes with exclude_unset to match the\nhistorical wire format); ``options`` is null for non-enum items.",
    "properties": {
     "configured": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Configured"
     },
     "default": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "string"
       },
       {
        "type": "boolean"
       }
      ],
      "title": "Default"
     },
     "description": {
      "title": "Description",
      "type": "string"
     },
     "editable": {
      "title": "Editable",
      "type": "boolean"
     },
     "group": {
      "title": "Group",
      "type": "string"
     },
     "key": {
      "title": "Key",
      "type": "string"
     },
     "label": {
      "title": "Label",
      "type": "string"
     },
     "options": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Options"
     },
     "restartRequired": {
      "title": "Restartrequired",
      "type": "boolean"
     },
     "secret": {
      "title": "Secret",
      "type": "boolean"
     },
     "type": {
      "enum": [
       "int",
       "bool",
       "string",
       "enum",
       "secret"
      ],
      "title": "Type",
      "type": "string"
     },
     "value": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "string"
       },
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Value"
     }
    },
    "required": [
     "key",
     "label",
     "description",
     "group",
     "type",
     "default",
     "editable",
     "secret",
     "restartRequired"
    ],
    "title": "RssHubConfigItem",
    "type": "object"
   },
   "RssHubConfigPatch": {
    "description": "PATCH /api/v1/rsshub/config body: allow-listed non-secret values.",
    "properties": {
     "values": {
      "additionalProperties": true,
      "title": "Values",
      "type": "object"
     }
    },
    "title": "RssHubConfigPatch",
    "type": "object"
   },
   "RssHubConfigView": {
    "description": "GET/PATCH /api/v1/rsshub/config (secrets never echoed).",
    "properties": {
     "configured": {
      "title": "Configured",
      "type": "boolean"
     },
     "groups": {
      "items": {
       "$ref": "#/components/schemas/RssHubConfigGroup"
      },
      "title": "Groups",
      "type": "array"
     },
     "pendingCount": {
      "title": "Pendingcount",
      "type": "integer"
     },
     "pendingSecrets": {
      "title": "Pendingsecrets",
      "type": "boolean"
     },
     "schemaVersion": {
      "title": "Schemaversion",
      "type": "integer"
     }
    },
    "required": [
     "schemaVersion",
     "configured",
     "pendingCount",
     "pendingSecrets",
     "groups"
    ],
    "title": "RssHubConfigView",
    "type": "object"
   },
   "RssHubCredentialCreate": {
    "description": "POST /api/v1/rsshub/credentials body (value is write-only).",
    "properties": {
     "domain": {
      "maxLength": 253,
      "minLength": 1,
      "title": "Domain",
      "type": "string"
     },
     "envKey": {
      "maxLength": 64,
      "minLength": 1,
      "title": "Envkey",
      "type": "string"
     },
     "kind": {
      "enum": [
       "cookie",
       "token",
       "api_key",
       "bearer",
       "other"
      ],
      "title": "Kind",
      "type": "string"
     },
     "name": {
      "maxLength": 80,
      "minLength": 1,
      "title": "Name",
      "type": "string"
     },
     "route": {
      "default": "",
      "maxLength": 200,
      "title": "Route",
      "type": "string"
     },
     "value": {
      "maxLength": 10000,
      "minLength": 1,
      "title": "Value",
      "type": "string"
     }
    },
    "required": [
     "name",
     "domain",
     "envKey",
     "kind",
     "value"
    ],
    "title": "RssHubCredentialCreate",
    "type": "object"
   },
   "RssHubCredentialMetadataPatch": {
    "description": "PATCH /api/v1/rsshub/credentials/{id} body (metadata only).",
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "route": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Route"
     }
    },
    "title": "RssHubCredentialMetadataPatch",
    "type": "object"
   },
   "RssHubCredentialValuePut": {
    "description": "PUT /api/v1/rsshub/credentials/{id}/value body.",
    "properties": {
     "value": {
      "maxLength": 10000,
      "minLength": 1,
      "title": "Value",
      "type": "string"
     }
    },
    "required": [
     "value"
    ],
    "title": "RssHubCredentialValuePut",
    "type": "object"
   },
...
   "RssHubPreviewRequest": {
    "description": "POST /api/v1/rsshub/preview body (0014): route + parameter values.",
    "properties": {
     "params": {
      "additionalProperties": {
       "type": "string"
      },
      "title": "Params",
      "type": "object"
     },
     "routeId": {
      "minLength": 1,
      "title": "Routeid",
      "type": "string"
     }
    },
    "required": [
     "routeId"
    ],
    "title": "RssHubPreviewRequest",
    "type": "object"
   },
...
   "SecretValuePut": {
    "description": "Write-only secret body (shared by AI keys and RSSHub secrets).\n\n0021 hardening: bounded length (same bound as the RSSHub secret\nschema) and control-character rejection at INPUT time — a control\ncharacter in an API key is never legitimate, and storing one would\npoison env-file rendering / upstream calls until the value is\nreplaced.",
    "properties": {
     "value": {
      "maxLength": 10000,
      "minLength": 1,
      "title": "Value",
      "type": "string"
     }
    },
    "required": [
     "value"
    ],
    "title": "SecretValuePut",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "LumiRSS BFF",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/rsshub/config": {
   "get": {
    "operationId": "get_rsshub_config_api_v1_rsshub_config_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RssHubConfigView"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "Get Rsshub Config"
   },
   "patch": {
    "description": "Update allow-listed non-secret desired values (validated + typed).\n\nSecrets are never accepted here — use the secret endpoints. Saving only\nupdates the DESIRED config; the UI reports restartRequired honestly.",
    "operationId": "patch_rsshub_config_api_v1_rsshub_config_patch",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RssHubConfigPatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RssHubConfigView"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Patch Rsshub Config"
   }
  },
...
  "/api/v1/rsshub/config/secrets/{key}": {
   "delete": {
    "description": "Clear one secret (explicit action).",
    "operationId": "delete_rsshub_secret_api_v1_rsshub_config_secrets__key__delete",
    "parameters": [
     {
      "in": "path",
      "name": "key",
      "required": true,
      "schema": {
       "title": "Key",
       "type": "string"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Rsshub Secret"
   },
   "put": {
    "description": "Write one route credential / secret (write-only, never read back).",
    "operationId": "put_rsshub_secret_api_v1_rsshub_config_secrets__key__put",
    "parameters": [
     {
      "in": "path",
      "name": "key",
      "required": true,
      "schema": {
       "title": "Key",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SecretValuePut"
       }
      }
     },
     "required": true
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Put Rsshub Secret"
   }
  },
  "/api/v1/rsshub/credentials": {
   "get": {
    "description": "Custom site/route credentials (values are write-only; configured\nflags only). Adding one NEVER fabricates an RSSHub route.",
    "operationId": "list_rsshub_credentials_api_v1_rsshub_credentials_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "additionalProperties": true,
          "type": "object"
         },
         "title": "Response List Rsshub Credentials Api V1 Rsshub Credentials Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "List Rsshub Credentials"
   },
   "post": {
    "operationId": "create_rsshub_credential_api_v1_rsshub_credentials_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RssHubCredentialCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Create Rsshub Credential Api V1 Rsshub Credentials Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Rsshub Credential"
   }
  },
  "/api/v1/rsshub/credentials/{credential_id}": {
   "delete": {
    "operationId": "delete_rsshub_credential_api_v1_rsshub_credentials__credential_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "credential_id",
      "required": true,
      "schema": {
       "title": "Credential Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Rsshub Credential"
   },
   "patch": {
    "operationId": "patch_rsshub_credential_api_v1_rsshub_credentials__credential_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "credential_id",
      "required": true,
      "schema": {
       "title": "Credential Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RssHubCredentialMetadataPatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Patch Rsshub Credential Api V1 Rsshub Credentials  Credential Id  Patch",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Patch Rsshub Credential"
   }
  },
  "/api/v1/rsshub/credentials/{credential_id}/value": {
   "put": {
    "operationId": "put_rsshub_credential_value_api_v1_rsshub_credentials__credential_id__value_put",
    "parameters": [
     {
      "in": "path",
      "name": "credential_id",
      "required": true,
      "schema": {
       "title": "Credential Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RssHubCredentialValuePut"
       }
      }
     },
     "required": true
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Put Rsshub Credential Value"
   }
  },
...
  "/api/v1/rsshub/preview": {
   "post": {
    "description": "Preview one configured RSSHub route — NON-MUTATING.\n\nConstructs the path server-side (validated + encoded parameters),\nfetches the generated feed from the server-configured RSSHub\ninstance, parses offline and reads the subscription list for\nalreadySubscribed. The returned feedUrl is the FreshRSS-facing\nsubscription URL; subscribing is POST /api/v1/subscriptions (0013).",
    "operationId": "rsshub_preview_api_v1_rsshub_preview_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RssHubPreviewRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FeedPreviewResult"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Rsshub Preview"
   }
  },
...
 }
}